requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pytesseract>=0.3.10
pdf2image>=1.16.0
pillow>=10.0.0
opencv-python>=4.8.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-multipart>=0.0.20